Fixed logic error in synch.h, added more comments, minor tweaks to matrix files, added collision testing

+++ data_structures/synch.h
@@ -1,11 +1,10 @@
 #include "model.h"
 #include "matrix.h"
 #include "tetromino.h"
-#include "types.h"
+#include "../types.h"
 
 UINT16 can_move_down(Model *model);
 
-UINT16 check_collision(const Matrix *board, const Tetromino *piece, UINT16 offset_x, UINT16 offset_y);
+UINT16 check_collision(const Matrix *board, const Tetromino *piece, int offset_x, int offset_y);
 
 void handle_tick (Model *model);
-
